Mid-work on making Prover and Verifier completely functional, storing all state in the macro-generated module/struct
