+++ src/macros.rs
@@ -120,6 +120,8 @@ macro_rules! define_proof {
             use $crate::curve25519_dalek::ristretto::CompressedRistretto;
 
             use $crate::toolbox::prover::Prover;
+            use $crate::toolbox::prover::PointVar;
+            use $crate::toolbox::prover::ScalarVar;
             use $crate::toolbox::verifier::Verifier;
 
             use $crate::toolbox::SchnorrCS;
@@ -131,7 +133,21 @@ macro_rules! define_proof {
             pub use $crate::merlin::Transcript;
             pub use $crate::{CompactProof, BatchableProof, ProofError};
 
-            pub struct $proof_module_name {}
+            pub struct $proof_module_name<'a> {
+                transcript: &'a mut Transcript,
+                scalars: Vec<Option<Scalar>>,
+                points: Vec<RistrettoPoint>,
+                point_labels: Vec<&'static [u8]>,
+                constraints: Vec<(usize, PointVar, Vec<(ScalarVar, PointVar)>)>,
+                subroutines: Vec<Prover<'a>>,
+            
+                //internals
+                commitments: Vec<CompressedRistretto>,
+                blindings: Vec<Option<Scalar>>,
+                fake_responses: Vec<Option<Scalar>>,
+                known_chal_shares: Vec<Option<Scalar>>,
+                pub nr_clauses: usize,
+            }
 
             /// The proof label committed to the transcript as a domain separator.
             pub const PROOF_LABEL: &'static str = $proof_label_string;
@@ -233,11 +249,23 @@ macro_rules! define_proof {
                 $(pub $common_var: CompressedRistretto,)*
             }
 
-            pub fn new() -> $proof_module_name {
-                $proof_module_name{}
+            pub fn new(transcript: &mut Transcript) -> $proof_module_name {
+                $proof_module_name{
+                    transcript,
+                    scalars: Vec::default(),
+                    points: Vec::default(),
+                    point_labels: Vec::default(),
+                    constraints: Vec::default(),
+                    subroutines: Vec::default(),
+                    commitments: Vec::default(),
+                    blindings: Vec::default(),
+                    fake_responses: Vec::default(),
+                    known_chal_shares: Vec::default(),
+                    nr_clauses: 0,
+                }
             }
 
-            impl $proof_module_name {
+            impl<'a> $proof_module_name<'a> {
 
                 /// The underlying proof statement generated by the macro invocation.
                 ///
@@ -274,7 +302,7 @@ macro_rules! define_proof {
                     clause_nr - 1
                 }
 
-                fn build_prover<'a>(
+                fn build_prover(
                     &self,
                     transcript: &'a mut Transcript,
                     assignments: ProveAssignments,
@@ -376,7 +404,7 @@ macro_rules! define_proof {
                     }, compressed))
                 }
 
-                fn build_verifier<'a>(
+                fn build_verifier(
                     &self,
                     transcript: &'a mut Transcript,
                     assignments: VerifyAssignments,
@@ -479,30 +507,303 @@ macro_rules! define_proof {
                 }
             }
 
-            // impl<A> IsSigmaProtocol<A> for $proof_module_name
-            // where
-            //     A: IsProverAssignments {
-            //     type Proof = BatchableProof;
-
-            //     fn commit(&mut self, assignments: A) -> Result<(), ProofError> {
-            //         Ok(())
-            //     }
-
-            //     fn challenge(&mut self) -> Scalar {
-            //         Scalar::zero()
-            //     }
-
-            //     fn response(&mut self, challenge: Scalar, assignments: A) -> Self::Proof {
-            //         BatchableProof{
-            //             challenges: vec![],
-            //             commitments: vec![],
-            //             responses: vec![]
-            //         }
-            //     }
-            //     fn simulate(&mut self, assignments: A) -> (Vec<Scalar>, Vec<Scalar>) {
-            //         (vec![], vec![])
-            //     }
-            // }
+            impl<'a, P> IsSigmaProtocol<P> for $proof_module_name<'a> where P: IsProverAssignments {
+                type Proof = BatchableProof;        // TODO we need to make this configurable
+            
+                fn commit(&mut self, assignments: P) -> Result<(), ProofError> {
+                    // Construct a TranscriptRng
+                    let mut rng_builder = self.transcript.build_rng();
+                    for scalar in &self.scalars {
+                        if scalar.is_some() {
+                            rng_builder = rng_builder.rekey_with_witness_bytes(b"", scalar.unwrap().as_bytes());
+                        }
+                    }
+                    let mut transcript_rng = rng_builder.finalize(&mut thread_rng());
+            
+                    // Generate a blinding factor for each secret variable
+                    let blindings = assignments.secret_vars().iter().map(|scalar| {
+                            return match scalar.is_some() {
+                                true => Some(Scalar::random(&mut transcript_rng)),
+                                false => None
+                            };
+                        })
+                        .collect::<Vec<Option<Scalar>>>();
+            
+                    // Commit to each blinded LHS
+                    let mut commitments = Vec::with_capacity(self.constraints.len());
+                    let mut fake_responses = Vec::with_capacity(self.constraints.iter().map(|cs| &cs.1).count());
+                    let mut shares = Vec::with_capacity(self.constraints.len());
+                    let mut clause_tracker = vec![None; self.nr_clauses+1];
+            
+                    match self.proof_type {
+                        ProofType::Shamir => {
+                            for (clause_nr, lhs_var, rhs_lin_combo) in &self.constraints {
+                                debug!("Constraint clause {}: {} = {}",
+                                    clause_nr,
+                                    str::from_utf8(self.point_labels[lhs_var.0]).unwrap(),
+                                    rhs_lin_combo.iter().map(
+                                        |(scal, pt)| format!("{} ^ {:?}", str::from_utf8(self.point_labels[pt.0]).unwrap(), scal.0)
+                                    ).collect::<Vec<String>>().join(" * ")
+                                );
+                    
+                                // The first [0] picks the first entry in the linear combination.  We can test against only this one because we
+                                // know that all entries in a compound && clause (represented by a single constraint) MUST either be all known, or
+                                // all unknown.  Though, we should have a way of checking this in the code before execution reaches this point.
+                                // The second .0 pulls out the ScalarVar which the PointVar is to be raised to.  The third .0 selects the single entry
+                                // of the ScalarVar tuple, which is a usize representing the index of the ScalarVar's value in self.scalars and self.blindings.
+                                let sv_index = (rhs_lin_combo[0].0).0;
+                                let commitment = match blindings[sv_index].is_some() {
+                                    true => {               // if we have a blinding, that means we have a secret value for this variable
+                                        clause_tracker[*clause_nr] = match clause_tracker[*clause_nr] {
+                                            // this is the first time we've worked with this clause, initialize the tracker
+                                            None => Some((1, 1)),
+                    
+                                            // We're adding a new known value, so increment the tracker
+                                            Some((terms, filled_in)) => Some((terms+1, filled_in+1)),
+                                        };
+                    
+                                        shares.push(None);
+                                        RistrettoPoint::multiscalar_mul(
+                                            rhs_lin_combo.iter().map(|(sc_var, _pt_var)| {
+                                                fake_responses.push(None);
+                                                blindings[sc_var.0].unwrap()
+                                            }),
+                                            rhs_lin_combo.iter().map(|(_sc_var, pt_var)| {
+                                                self.points[pt_var.0]
+                                            }),
+                                        )
+                                    }
+                                    false => {              // if we don't have a blinding, we don't have a secret value and will be faking one
+                                        clause_tracker[*clause_nr] = match clause_tracker[*clause_nr] {
+                                            // this is the first time we've worked with this clause, initialize the tracker
+                                            None => Some((1, 0)),
+                    
+                                            // we've added no new information here, so just count a new clause
+                                            Some((terms, filled_in)) => Some((terms+1, filled_in)),
+                                        };
+                    
+                                        let challenge = Scalar::random(&mut transcript_rng);
+                                        shares.push(Some(challenge));
+                                        RistrettoPoint::multiscalar_mul(
+                                            rhs_lin_combo.iter()
+                                                .map(|(_sc_var, _pt_var)| {
+                                                    let response = Scalar::random(&mut transcript_rng);
+                                                    fake_responses.push(Some(response));
+                                                    response
+                                                })
+                                                .chain(iter::once(-&challenge)),
+                                            rhs_lin_combo.iter()
+                                                .map(|(_sc_var, pt_var)| self.points[pt_var.0])
+                                                .chain(iter::once(self.points[lhs_var.0])),
+                                        )
+                                    }
+                                };
+                    
+                                let encoding = self
+                                    .transcript
+                                    .append_blinding_commitment(self.point_labels[lhs_var.0], &commitment);
+            
+                                trace!("Appending blinding {:?}", encoding);
+                    
+                                commitments.push(encoding);
+                            }
+                            // Now that we know all the clauses, make sure at least one is full
+                            let mut have_one = false;
+                            for (i, clause) in clause_tracker.iter().enumerate() {
+                                match clause {
+                                    Some((terms, filled_in)) => {
+                                        debug!("Clause {} has {} terms and we know {} keys", i, terms, filled_in);
+                                        if terms == filled_in { have_one = true; break; }
+                                    },
+                                    None => (),
+                                }
+                            }
+                            if !have_one {
+                                return Err(ProofError::InsufficientKeys)
+                            }
+                        },
+                        ProofType::Xor => {
+                            // General approach: we're going to have C clauses, each of which will have D constraints, each of which might have E terms
+                            // We've already generated a blinding for each possible E term
+                            // We need to track the D constraints so we can calculate completeness of each C
+                            // We should only have one fully-complete C, for which the c_i value can be calculated
+                            // For all other Cs except 1, we can randomize the c_i value
+            
+                            // do ClauseTracker logic first, so we can trigger on it later on
+                            for (clause_nr, lhs_var, rhs_lin_combo) in &self.constraints {
+                                debug!("Constraint clause {}: {} = {}",
+                                    clause_nr,
+                                    str::from_utf8(self.point_labels[lhs_var.0]).unwrap(),
+                                    rhs_lin_combo.iter().map(
+                                        |(scal, pt)| format!("{} ^ {:?}", str::from_utf8(self.point_labels[pt.0]).unwrap(), scal.0)
+                                    ).collect::<Vec<String>>().join(" * ")
+                                );
+            
+                                let sv_index = (rhs_lin_combo[0].0).0;
+                                match blindings[sv_index].is_some() {
+                                    true => {               // if we have a blinding, that means we have a secret value for this variable
+                                        clause_tracker[*clause_nr] = match clause_tracker[*clause_nr] {
+                                            // this is the first time we've worked with this clause, initialize the tracker
+                                            None => Some((1, 1)),
+                    
+                                            // We're adding a new known value, so increment the tracker
+                                            Some((terms, filled_in)) => Some((terms+1, filled_in+1)),
+                                        };
+                                    },
+                                    false => {
+                                        clause_tracker[*clause_nr] = match clause_tracker[*clause_nr] {
+                                            // this is the first time we've worked with this clause, initialize the tracker
+                                            None => Some((1, 0)),
+                    
+                                            // we've added no new information here, so just count a new clause
+                                            Some((terms, filled_in)) => Some((terms+1, filled_in)),
+                                        };
+                                    }
+                                };
+                            };
+            
+                            // Now that we know all the clauses, make sure at least one is full, and generate one challenge share
+                            // per clause (NOT per constraint!)
+                            let mut have_one = false;
+                            for (i, clause) in clause_tracker.iter().enumerate() {
+                                match clause {
+                                    Some((terms, filled_in)) => {
+                                        debug!("Clause {} has {} terms and we know {} keys", i, terms, filled_in);
+                                        if terms == filled_in {
+                                            // this is a fully-known clause, so we'll be able to calculate the challenge later on
+                                            have_one = true;
+                                            shares.push(None);
+                                        } else {
+                                            // this is a faked clause, so we need to randomize a challenge for it
+                                            shares.push(Some(random_scalar(&mut transcript_rng)));
+                                        };
+                                    },
+                                    None => (),
+                                }
+                            }
+                            if !have_one {
+                                return Err(ProofError::InsufficientKeys)
+                            }
+            
+                            // Using the shares, calculate our commitments
+                            for (clause_nr, lhs_var, rhs_lin_combo) in &self.constraints {
+                                let sv_index = (rhs_lin_combo[0].0).0;
+                                let commitment = match blindings[sv_index].is_some() {
+                                    true => {               // if we have a blinding, that means we have a secret value for this variable        
+                                        RistrettoPoint::multiscalar_mul(
+                                            rhs_lin_combo.iter().map(|(sc_var, _pt_var)| {
+                                                fake_responses.push(None);
+                                                blindings[sc_var.0].unwrap()
+                                            }),
+                                            rhs_lin_combo.iter().map(|(_sc_var, pt_var)| {
+                                                self.points[pt_var.0]
+                                            }),
+                                        )
+                                    }
+                                    false => {              // if we don't have a blinding, we don't have a secret value and will be faking one        
+                                        let challenge = shares[*clause_nr - 1].unwrap();
+                                        RistrettoPoint::multiscalar_mul(
+                                            rhs_lin_combo.iter()
+                                                .map(|(_sc_var, _pt_var)| {
+                                                    let response = Scalar::random(&mut transcript_rng);
+                                                    fake_responses.push(Some(response));
+                                                    response
+                                                })
+                                                .chain(iter::once(-&challenge)),
+                                            rhs_lin_combo.iter()
+                                                .map(|(_sc_var, pt_var)| self.points[pt_var.0])
+                                                .chain(iter::once(self.points[lhs_var.0])),
+                                        )
+                                    }
+                                };
+                    
+                                let encoding = self
+                                    .transcript
+                                    .append_blinding_commitment(self.point_labels[lhs_var.0], &commitment);
+                                
+                                trace!("Appending blinding {:?}", encoding);
+                    
+                                commitments.push(encoding);
+                            }
+                        },
+                        _ => {
+                            panic!("This isn't implemented yet!");
+                        }
+                    }
+            
+                    debug!("Prover generated shares {:?}", shares);
+                    
+                    self.blindings = blindings;
+                    self.commitments = commitments;
+                    self.fake_responses = fake_responses;
+                    self.known_chal_shares = shares;
+                    Ok(())
+                }
+            
+                fn simulate(&mut self, _assignments: P) -> (Vec<Scalar>, Vec<Scalar>) {
+                    (vec![], vec![])
+                    // TODO
+                }
+            
+                fn challenge(&mut self) -> Scalar {
+                    let challenge = self.transcript.get_challenge(b"chal");
+                    debug!("Prover's transcript challenge is {:?}", challenge);
+                    challenge
+                }
+            
+                // TODO it looks like we don't modify the transcript here; is that what we want?
+                fn response(&mut self, challenge: Scalar, _assignments: P) -> Self::Proof {
+                    let blindings = &self.blindings;
+                    let fake_responses = &self.fake_responses;
+                    let responses: Vec<Scalar>;
+            
+                    let mut rng = rand::thread_rng();
+            
+                    let mut sharing: Box<dyn SecretSharing> = match self.proof_type {
+                        ProofType::Xor => {
+                            Box::from(Xor::new(&mut rng)) as Box<dyn SecretSharing>
+                        },
+                        ProofType::Shamir => {
+                            let threshold = 1 + self.known_chal_shares.iter().filter(|r| r.is_some()).count();
+                            Box::from(Shamir::new(threshold, &mut rng)) as Box<dyn SecretSharing>
+                        },
+                        _ => {
+                            panic!("This isn't implemented yet!");
+                        }
+                    };
+            
+                    let mut challenges = sharing.complete(&challenge, &self.known_chal_shares).unwrap();
+                    trace!("Challenges: {:?}", challenges);
+            
+                    match self.proof_type {
+                        ProofType::Xor => {
+                            challenges = self.constraints.iter().map(|(clause_nr, _, _)| {
+                                challenges[*clause_nr - 1]
+                            }).collect();
+                        },
+                        _ => ()
+                    };
+            
+                    responses = self.scalars.iter().zip(blindings)
+                                .zip(fake_responses)
+                                .zip(&challenges)
+                                .map(| (((scalar, blinding), fake_response), challenge) | {
+                                    match fake_response.is_some() {
+                                        true => fake_response.unwrap(),
+                                        false => scalar.unwrap() * challenge + blinding.unwrap(),
+                                    }
+                                })
+                                .collect::<Vec<Scalar>>();
+            
+                    let out_shares = challenges;
+                    let commitments = self.commitments.clone();
+                    Self::Proof{
+                        challenges: out_shares,
+                        responses,
+                        commitments,
+                    }
+                }
+            }
 
             // #[cfg(all(feature = "bench", test))]
             // mod bench {
